+++ packages/e2e/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cogitator-ai/e2e
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @cogitator-ai/deploy@0.1.2
+  - @cogitator-ai/cli@0.3.4
+
 ## 0.1.7
 
 ### Patch Changes